% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buffer.percentage}
\alias{buffer.percentage}
\title{The percentage to buffer shapefiles in the LandUnavailabilityR Package}
\format{
A numeric vector of length 1 with the buffer percentage
}
\usage{
buffer.percentage
}
\description{
The percentage to buffer shapefiles in the LandUnavailabilityR Package
}
\keyword{datasets}
