% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildableLand.R
\docType{class}
\name{BuildableLand}
\alias{BuildableLand}
\title{Class to calculate the amount of buildable land (meters) in a polygon region}
\description{
Calculate the amount of buildable land in a region
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BuildableLand-finish}{\code{BuildableLand$finish()}}
\item \href{#method-BuildableLand-calc_buildable_land}{\code{BuildableLand$calc_buildable_land()}}
\item \href{#method-BuildableLand-new}{\code{BuildableLand$new()}}
\item \href{#method-BuildableLand-clone}{\code{BuildableLand$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BuildableLand-finish"></a>}}
\if{latex}{\out{\hypertarget{method-BuildableLand-finish}{}}}
\subsection{Method \code{finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildableLand$finish()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BuildableLand-calc_buildable_land"></a>}}
\if{latex}{\out{\hypertarget{method-BuildableLand-calc_buildable_land}{}}}
\subsection{Method \code{calc_buildable_land()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildableLand$calc_buildable_land()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BuildableLand-new"></a>}}
\if{latex}{\out{\hypertarget{method-BuildableLand-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildableLand$new(
  slope.tif.path,
  mask.tif.path,
  slope.catalog,
  mask.catalog,
  shapefile,
  buffer.meters = NULL,
  buffer.percent = NULL,
  raster.crs = NULL,
  parallel.cores = (detectCores(logical = FALSE)),
  safe.compute = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BuildableLand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BuildableLand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildableLand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
