% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land.unavailability.example}
\alias{land.unavailability.example}
\title{An example of land unavailability output for saiz 1999 MSAs}
\format{
a \code{data.table} with the saiz land unavailability output
}
\usage{
land.unavailability.example
}
\description{
An example of land unavailability output for saiz 1999 MSAs
}
\keyword{datasets}
