% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandUnavailability.R
\docType{class}
\name{LandUnavailability}
\alias{LandUnavailability}
\title{Class to calculate the the percentage of land unavailability given
tif files and usgs catalog}
\format{
\code{\link{R6Class}} object.

#' @section Arguments for the \code{new} method

\describe{
  \item{\code{type}}{The type of analysis. The options are \code{"slope"} or \code{"water"}}
  \item{\code{raster.files.path}}{The path to the .tif files}
  \item{\code{raster.catalog}}{A \code{data.table} or file path to an rds file with the USGS or raster catalog}
  \item{\code{shapefile}}{A \code{sf} object or a path to a shapefile.}
  \item{\code{buffer.meters = NULL}}{How many meters to buffer each shapefile. The default is \code{NULL} for no buffer in meters}
  \item{\code{buffer.percent = NULL}}{What percentage to buffer each shapefile. Typically between 0 and 1. The default is \code{NULL} for no buffer in percentage terms.}
  \item{\code{parallel.cores = detectCores(logical = FALSE)}}{Number of parallel cores. Set to zero for sequential computation}
  \item{\code{safe.compute = FALSE}}{A safe compute option that will process each raster individually. It's slower, but likely process the rasters successfully}
}

#' @section Public (Mutable) Attributes.

\describe{
   \item{\code{percent.unvailable}}{A \code{data.table} with the \code{GEOID} and the percentage of land unavailable. If the \code{NA} values correspond to an error for a given \code{GEOID}}
}

#' @section Public Methods.

\describe{
  \item{\code{new(type = c("slope", "water"), raster.files.path, raster.catalog, shapefile, buffer.meters = NULL, buffer.percent = NULL, parallel.cores = detetCores(logical = FALSE), safe.compute = FALSE)}}{Instantiate an of the LandUnavailability class with \code{LandUnavailability$new()} function.}
   \item{\code{calc_unavailable_land}}{To calculate the percentage of unavailable land}
   \item{\code{finish()}}{To delete any temporary files associated with computation}
}

#' @section Locked Data Attributes (Active Bindings. View Only, But Can Assigned to Other Objects):

\describe{
   \item{\code{type}}{(character) The type of analysis. Either "slope" or "water"}
   \item{\code{id}}{(character) The ID used for saving files for this object}
   \item{\code{temp.geoid.folder}}{(character) The temporary folder where we save the land unavailability for each GEOID}
   \item{\code{raster.files.path}}{The path to the tif files}
   \item{\code{raster.catalog}}{A \code{data.table} or path to an RDS file with a \code{data.table} with the catalog of USGS or raster files}
   \item{\code{shapefile}}{A simple features \code{sf} shapefile polygon or the path to a shapefile to read in by \code{sf}}
   \item{\code{raster.list}}{The list of raster objects}
   \item{\code{raster.crs}}{The CRS for the raster objects. This will vary depending if the unavailability analysis is for \code{"slope"} or \code{"water"}}
   \item{\code{sf.polygon}}{The \code{sf} polygon object}
   \item{\code{buffer.meters}}{The buffer meters. If \code{NULL}, then no buffer in meters is applied}
   \item{\code{buffer.percent}}{The buffer; typically between 0 and 1. If \code{NULL}, then no buffer in percentage terms is applied}
   \item{\code{polygon.raster.intersection}}{A \code{data.table} with the USGS IDs for each GEOID}
}
}
\value{
Object \code{\link{R6Class}} with methods accessing land
    unavailability data
}
\description{
Calculate the percentage of land unavailability using \code{out <-
LandUnavailability$new(type, raster.files.path, raster.catalog, shapefile, buffer.meters
= NULL, buffer.percent = NULL, parallel.cores = detectCores(logical = FALSE),
safe.compute = FALSE)}
}
\details{
The function will generate temporary files that can be used in case
of interuption of computatation. The names of temporary files will
be based on the \code{id} (see private methods below) and the geoid
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LandUnavailability-finish}{\code{LandUnavailability$finish()}}
\item \href{#method-LandUnavailability-calc_unavailable_land}{\code{LandUnavailability$calc_unavailable_land()}}
\item \href{#method-LandUnavailability-new}{\code{LandUnavailability$new()}}
\item \href{#method-LandUnavailability-clone}{\code{LandUnavailability$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandUnavailability-finish"></a>}}
\if{latex}{\out{\hypertarget{method-LandUnavailability-finish}{}}}
\subsection{Method \code{finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandUnavailability$finish()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandUnavailability-calc_unavailable_land"></a>}}
\if{latex}{\out{\hypertarget{method-LandUnavailability-calc_unavailable_land}{}}}
\subsection{Method \code{calc_unavailable_land()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandUnavailability$calc_unavailable_land()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandUnavailability-new"></a>}}
\if{latex}{\out{\hypertarget{method-LandUnavailability-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandUnavailability$new(
  type = c("slope", "water"),
  raster.files.path,
  raster.catalog,
  shapefile,
  buffer.meters = NULL,
  buffer.percent = NULL,
  raster.crs = NULL,
  fun.unavail.raster = NULL,
  parallel.cores = (detectCores(logical = FALSE)),
  safe.compute = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandUnavailability-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LandUnavailability-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandUnavailability$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
