% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandUnavailabilityR_main.R
\docType{package}
\name{LandUnavailabilityR-package}
\alias{LandUnavailabilityR}
\alias{LandUnavailabilityR-package}
\title{Functions and an \code{R6} class for Land Unavailability}
\description{
Compute Land Unavailability from Slope and Land Use GIS Tif Files.
}
\author{
\strong{Maintainer}: Chandler Lutz \email{chandler.lutz@gmail.com}

}
