% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cleaning_functions.R
\name{lu_calc_slope_unavailable}
\alias{lu_calc_slope_unavailable}
\title{Calculate the slope unavailability from a raster object}
\usage{
lu_calc_slope_unavailable(raster.values, slope.degrees.cutoff = 8.531)
}
\arguments{
\item{raster.values}{the values from a \code{raster} object
with the slope calculated from the \code{raster::slope}
function}

\item{slope.degrees.cutoff}{(numeric) The cutoff for a steep slope
in degrees. The default is \code{8.531} which corresponds to
the 15 percent used by Saiz (2010).}
}
\value{
(data.table) the percentage of land between 0 and 100 that
    is unavailable due to a steep slope
}
\description{
Note: this function will not check if the raster slope has been
calculated. See \url{https://rpubs.com/ajlyons/rspatialdata} for an
overview.
}
\details{
Also note that degrees can be calculated from the slope (in
percent) using the inverse tangent. See
\url{https://www.archtoolbox.com/measurements/geometry/slope.html}
}
