% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cleaning_functions.R
\name{lu_calc_usgs_water_unavailable}
\alias{lu_calc_usgs_water_unavailable}
\title{Calculate the percentage of water unavailability from a usgs raster
object}
\usage{
lu_calc_usgs_water_unavailable(raster.values)
}
\arguments{
\item{raster.values}{The values from a \code{raster}}
}
\value{
(data.table) the percentage of land between 0 and 100 that
    is unavailable due to a (1) water and (2) wetlands
}
\description{
Note, this object will rely on Land Cover Data and calculate
the percentage of land unavailable due to water and wetlands.  The
relevant USGS codes are as follows: water -- 11; wetlands 90 and 95
}
