% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cleaning_functions.R
\name{lu_check_raster_error}
\alias{lu_check_raster_error}
\title{Check for a raster error}
\usage{
lu_check_raster_error(file, plot.check = FALSE)
}
\arguments{
\item{file}{(string) a path to a tif object that will be read in by
\code{raster}}

\item{plot.check}{(logical) If set to \code{TRUE}, the function
will check if the raster can be plotted. This is a deeper check
of raster functionality}
}
\value{
if there is an error in reading or plotting the raster from
    \code{file}, the function will return \code{file}. Otherwise it
    will return \code{NULL}
}
\description{
Check for a raster error
}
