% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census_shp_files.R
\name{lu_download_census_shp_pre2010}
\alias{lu_download_census_shp_pre2010}
\title{Download pre-2010 US Census Shp files and convert to sf objects}
\usage{
lu_download_census_shp_pre2010(
  url,
  base.download.url,
  xpath = "//*[@id=\\"innerPage\\"]/table",
  testing.files = NULL
)
}
\arguments{
\item{url}{the URL where the list of files shp files can be found
on the census website}

\item{base.download.url}{the base url from where we can download a
zipped pre-2010 zipped shape file from the US census}

\item{xpath}{the \code{rvest} \code{xpath}.}

\item{testing.files}{a \code{numeric} vector with the files that
will be used for testing. Set to \code{NULL} for
production. The default is \code{NULL}. If \code{NULL}, the
function will process all files}
}
\value{
a \code{list} with all sf objects for the shapefiles. Note: for
    some sets of census shapefiles, the names of the sf objets can
    differ.
}
\description{
Census shape files from https://www.census.gov/geo/maps-data/data/prev_cartbndry_names.html
}
\examples{
## for zip3
url <- "https://www2.census.gov/geo/tiger/PREVGENZ/zt/z300shp/"
base.download.url <- "https://www2.census.gov/geo/tiger/PREVGENZ/zt/z300shp/"
\dontrun{
lu_download_census_shp_pre2010(url, base.download.url)
}
}
