% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cleaning_functions.R
\name{lu_get_raster_crs}
\alias{lu_get_raster_crs}
\title{The CRS of the raster files}
\usage{
lu_get_raster_crs(raster.file.path)
}
\arguments{
\item{raster.file.path}{(string) The path to the raster TIF
files}
}
\value{
(string) The CRS for the
}
\description{
Note: This function will return the CRS for the first file in
\code{raster.file.path}
}
