% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cleaning_functions.R
\name{lu_get_raster_list}
\alias{lu_get_raster_list}
\title{Get a list of raster files}
\usage{
lu_get_raster_list(raster.file.path)
}
\arguments{
\item{raster.file.path}{(string) The path to the raster TIF
files}
}
\value{
A list where each element of the list contains a
    \code{raster} object for the each raster tif file
}
\description{
Get a list of raster files
}
