% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_cleaning_functions.R
\name{lu_intersect_polygon_extent_raster}
\alias{lu_intersect_polygon_extent_raster}
\title{Get the raster IDs that intersect with a single polygon extent}
\usage{
lu_intersect_polygon_extent_raster(polygon.extent, raster.catalog)
}
\arguments{
\item{polygon.extent}{An \code{Extent} object created by
\code{raster::extent} for a polygon. We want to find the raster
IDs that intersect with \code{polygon.extent}. Note that this
is an \code{Extent} object for a single polygon}

\item{raster.catalog}{A \code{data.table} with the raster. For the
US, this will be the USGS catalog of raster objects. Note:
there needs to be a column titled "raster.extent" of class
\code{Extent} in and another column title "sourceId" with the
source IDs for the Raster Files}
}
\value{
A \code{character} vector with with the raster IDs that
    intersect polygon.extent
}
\description{
Get the raster IDs that intersect with a single polygon extent
}
