% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_cleaning_functions.R
\name{lu_intersect_sf_raster}
\alias{lu_intersect_sf_raster}
\title{Intersect a \code{sf} (simple features) object and a raster catalog}
\usage{
lu_intersect_sf_raster(sf.obj, raster.catalog)
}
\arguments{
\item{sf.obj}{A \code{sf} (simple features) object}

\item{raster.catalog}{A raster catalog that contains a \code{Extent}
object created by \code{raster::extent()}}
}
\value{
A \code{data.table} with with the boundry box for each
    polygon in \code{sf.obj} and the corresponing raster source IDs
    that intersect with each polygon in \code{sf.obj}
}
\description{
This function will return a \code{data.table} with the boundry box
and the raster IDs that intersect with a \code{sf} object.
}
