% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_cleaning_functions.R
\name{lu_sf_bbox}
\alias{lu_sf_bbox}
\title{Get the bounding box (bbox) for a \code{sf} (simple features) object}
\usage{
lu_sf_bbox(sf.obj)
}
\arguments{
\item{sf.obj}{A \code{sf} (simple features) object. \code{sf.obj}
must have a column named "GEOID" with the geographic identifier
for each polygon}
}
\value{
a \code{data.table} with the bouding box and the following
    columns (in order): \code{GEOID}, \code{xmin}, \code{xmax},
    \code{ymin}, \code{ymax}
}
\description{
Get the bounding box (bbox) for a \code{sf} (simple features) object
}
