% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_cleaning_functions.R
\name{lu_st_transform_all}
\alias{lu_st_transform_all}
\title{Transform ALL simple feature columns (sfc) to a new CRS}
\usage{
lu_st_transform_all(sf.obj, crs, ...)
}
\arguments{
\item{sf.obj}{A \code{sf} (simple features) polygon object}

\item{crs}{A valid CRS}

\item{...}{Other arguments to be passed to \code{st_transform}}
}
\value{
a \code{sf} object where of the simple feature columns are
    transformed to the new CRS
}
\description{
This function is useful when a simple features object has multiple
simple feature (polygon columns)
}
