% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cleaning_functions.R
\name{lu_total_land_unavailability}
\alias{lu_total_land_unavailability}
\title{Calculate the Total unavailability from a Slope and Water from a
\code{LandUnavailability} R6 instance}
\usage{
lu_total_land_unavailability(LU1, LU2, polygon.name.col = NULL)
}
\arguments{
\item{LU1}{An instance from the LandUnavailability R6 Class}

\item{LU2}{An instance from the LandUnavailability R6 Class}

\item{polygon.name.col}{(string) the name of the column name with a
polygon names}
}
\value{
a \code{data.table} with the total land unavailability
    output
}
\description{
Calculate the Total unavailability from a Slope and Water from a
\code{LandUnavailability} R6 instance
}
