% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saiz.sf}
\alias{saiz.sf}
\title{The \code{sf} Saiz shapefile for 1999 MSAs}
\format{
an \code{sf} object with the saiz shapefiles
}
\usage{
saiz.sf
}
\description{
The \code{sf} Saiz shapefile for 1999 MSAs
}
\keyword{datasets}
